requirements.txt
faker==23.2.0
python-dateutil
confluent-kafka==2.2.0